# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Statsio API
  version: 1.0.0
  description: API documentation for the Statsio backend.
servers:
  - url: http://localhost:8080
    description: Local Docker environment
...
paths:
...
  /api/statsdata/{id}:
    get:
      tags: [StatsData]
      summary: Get a StatsData document (blocs + dataSources) pour édition
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/StatsDataDocumentId'
      responses:
        '200':
          description: Document trouvé
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/StatsDataDocumentPayload'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [StatsData]
      summary: Update a StatsData document (champs partiels ou document complet)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/StatsDataDocumentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatsDataDocumentPatch'
      responses:
        '200':
          description: Document updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsDataDocumentEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags: [StatsData]
      summary: Delete a StatsData document
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/StatsDataDocumentId'
      responses:
        '204':
          description: Deleted (no body)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/statsdata/{documentId}/sources/{sourceId}:
    get:
      tags: [StatsData]
      summary: Get a data source by id
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/StatsDataDocumentIdPath'
        - $ref: '#/components/parameters/StatsDataSourceId'
      responses:
        '200':
          description: Source found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsDataSourceEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [StatsData]
      summary: Update a data source
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/StatsDataDocumentIdPath'
        - $ref: '#/components/parameters/StatsDataSourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatsDataSourcePatch'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/StatsDataSourcePatchMultipart'
      responses:
        '200':
          description: Source updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsDataSourceEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags: [StatsData]
      summary: Delete a data source
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/StatsDataDocumentIdPath'
        - $ref: '#/components/parameters/StatsDataSourceId'
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
  parameters:
...
    StatsDataDocumentId:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
    StatsDataDocumentIdPath:
      in: path
      name: documentId
      required: true
      schema:
        type: string
        format: uuid
    StatsDataSourceId:
      in: path
      name: sourceId
      required: true
      schema:
        type: string
        format: uuid
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    StatsDataDocumentPatch:
      type: object
      properties:
        title:
          type: string
        subtitle:
          type: string
        visibility:
          type: string
          enum: [private, team, public]
        blocks:
          type: array
          items:
            type: object
            additionalProperties: true
    StatsDataDocumentPayload:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        subtitle:
          type: string
        visibility:
          type: string
          enum: [private, team, public]
        blocks:
          type: array
          items:
            type: object
            additionalProperties: true
        dataSources:
          type: array
          description: Sources liées (GET /sources ou inclus après create/update document)
          items:
            type: object
            additionalProperties: true
        slug:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    StatsDataSourcePatch:
      type: object
      properties:
        name:
          type: string
        sortOrder:
          type: integer
        manualData:
          type: object
          additionalProperties: true
        apiUrl:
          type: string
          format: uri
        apiKey:
          type: string
          nullable: true
        verify:
          type: boolean
        responseRoot:
          type: string
          nullable: true
        normalizationMapping:
          $ref: '#/components/schemas/StatsDataNormalizationMapping'
    StatsDataSourcePatchMultipart:
      allOf:
        - $ref: '#/components/schemas/StatsDataSourcePatch'
        - type: object
          properties:
            file:
              type: string
              format: binary
    StatsDataSourceEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          type: object
          additionalProperties: true
    StatsDataNormalizationMapping:
      type: object
      description: |
        Mapping brut → lignes { keys, values }. Au moins keyFields ou valueFields non vide pour un refresh.
        Aucune convention imposée par fournisseur : pour chaque API, renseigner explicitement rowPath (liste de lignes dans la réponse parsée) et from (chemins dot par rapport à chaque ligne brute).
      properties:
        rowPath:
          type: string
          nullable: true
          description: Chemin dot vers la liste de lignes dans le JSON/XML parsé (omit ou null seulement si la racine parsée est déjà un tableau de lignes)
        keyFields:
          type: array
          items:
            type: object
            required: [name]
            properties:
              name:
                type: string
              from:
                type: string
                nullable: true
                description: Chemin dot dans la ligne brute ; défaut = name
        valueFields:
          type: array
          items:
            type: object
            required: [name]
            properties:
              name:
                type: string
              from:
                type: string
                nullable: true
        staticKeys:
          type: object
          additionalProperties: true
          description: Clés ajoutées à keys pour chaque ligne
...
    StatsDataDocumentEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          $ref: '#/components/schemas/StatsDataDocumentPayload'
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          nullable: true
        message:
          type: string
        error:
          type: string
          nullable: true
        data:
          nullable: true
    ValidationErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          nullable: true
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string